+++ src/components/authPage/layouts/authPageLayout/index.tsx
@@ -1,27 +1,21 @@
 import cn from '@/utils/cn';
 import { ChildrenProp } from '../../common.types';
 import { AUTH_PAGE_CONTENT_STYLES, AUTH_PAGE_LAYOUT_STYLES } from './constants';
-import dynamic from 'next/dynamic';
-const WarningToast = dynamic(() => import('../warningToast/index'), {
-  ssr: false,
-});
 
 export default function AuthPageLayout({
   children,
 }: ChildrenProp): JSX.Element {
   return (
     <div className={AUTH_PAGE_LAYOUT_STYLES}>
-      <WarningToast>
-        <div
-          className={cn(
-            AUTH_PAGE_CONTENT_STYLES,
-            'xl:max-w-[640px]',
-            'sm:max-w-[327px]',
-          )}
-        >
-          {children}
-        </div>
-      </WarningToast>
+      <div
+        className={cn(
+          AUTH_PAGE_CONTENT_STYLES,
+          'xl:max-w-[640px]',
+          'sm:max-w-[327px]',
+        )}
+      >
+        {children}
+      </div>
     </div>
   );
 }
