+++ src/constants/warningMessages.ts
@@ -1,0 +1,5 @@
+// key에 쿼리가 될 값을 입력하고 value에 표시할 메시지를 입력해주세요.
+export const WARNING_MESSAGES = {
+  customerAccountExist: '알빈 유저 계정이 이미 존재합니다.',
+  moverAccountExist: '기사님 계정이 이미 존재합니다.',
+} as const;
