+++ src/providers/_provider.tsx
@@ -1,12 +1,15 @@
 import { PropsWithChildren } from 'react';
 import QueryProvider from '@/providers/queryProvider';
 import ToasterProvider from '@/providers/toastProvider';
+import WarningProvider from './warningProvider';
 
 // provider 관련 코드 합치기
 export default function Provider({ children }: PropsWithChildren) {
   return (
     <QueryProvider>
-      <ToasterProvider>{children}</ToasterProvider>
+      <ToasterProvider>
+        <WarningProvider>{children}</WarningProvider>
+      </ToasterProvider>
     </QueryProvider>
   );
 }
